+++ Event/xAOD/xAODTrigL1Calo/xAODTrigL1Calo/versions/CPMTowerAuxContainer_v2.h
@@ -4,7 +4,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: CPMTowerAuxContainer_v2.h 642659 2015-01-29 12:41:06Z morrisj $
+// $Id: CPMTowerAuxContainer_v2.h 646335 2015-02-12 01:16:10Z morrisj $
 #ifndef XAODTRIGL1CALO_VERSIONS_CPMTOWERAUXCONTAINER_V2_H
 #define XAODTRIGL1CALO_VERSIONS_CPMTOWERAUXCONTAINER_V2_H
 
@@ -22,19 +22,19 @@ namespace xAOD{
   ///
   /// @author John Morris <john.morris@cern.ch>
   ///
-  /// $Revision: 642659 $
-  /// $Date: 2015-01-29 13:41:06 +0100 (Thu, 29 Jan 2015) $    
+  /// $Revision: 646335 $
+  /// $Date: 2015-02-12 02:16:10 +0100 (Thu, 12 Feb 2015) $    
 
   class CPMTowerAuxContainer_v2 : public AuxContainerBase {
   public:
     // Default constructor
-    CPMTowerAuxContainer_v2();
+    CPMTowerAuxContainer_v2();    
 
   private:
-    std::vector<std::vector<int> > emEnergyVec;
-    std::vector<std::vector<int> > hadEnergyVec;
-    std::vector<std::vector<int> > emErrorVec;
-    std::vector<std::vector<int> > hadErrorVec;
+    std::vector<std::vector<uint_least8_t> > emEnergyVec;
+    std::vector<std::vector<uint_least8_t> > hadEnergyVec;
+    std::vector<std::vector<uint_least8_t> > emErrorVec;
+    std::vector<std::vector<uint_least8_t> > hadErrorVec;
     std::vector<float> phi;
     std::vector<float> eta;
     std::vector<uint_least8_t> peak;
